More refactoring: rename package, move some traits
* cache.client -> cache.memcached
* Move all non-Memcached-specific stuff to parent cache package
* move tests to correct packages
