+++ app/com/m3/octoparts/cache/CacheOps.scala
@@ -1,4 +1,4 @@
-package com.m3.octoparts.cache.client
+package com.m3.octoparts.cache
 
 import com.m3.octoparts.cache.directive.CacheDirective
 import com.m3.octoparts.cache.versioning._
