+++ app/com/m3/octoparts/cache/memcached/InMemoryRawCache.scala
@@ -1,5 +1,6 @@
-package com.m3.octoparts.cache.client
+package com.m3.octoparts.cache.memcached
 
+import com.m3.octoparts.cache.RawCache
 import shade.inmemory.{ InMemoryCache => ShadeInMemoryCache }
 import shade.memcached.Codec
