+++ app/com/m3/octoparts/repository/MutableCachingRepository.scala
@@ -1,6 +1,6 @@
 package com.m3.octoparts.repository
 
-import com.m3.octoparts.cache.client.Cache
+import com.m3.octoparts.cache.Cache
 import com.m3.octoparts.cache.key.HttpPartConfigCacheKey
 import com.m3.octoparts.http.HttpClientPool
 import com.m3.octoparts.model.config.ConfigModel
