+++ app/com/m3/octoparts/repository/RepositoriesModule.scala
@@ -1,6 +1,6 @@
 package com.m3.octoparts.repository
 
-import com.m3.octoparts.cache.client.Cache
+import com.m3.octoparts.cache.Cache
 import com.m3.octoparts.http.HttpClientPool
 import scaldi.Module
